requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
mcrcon==0.7.0
python-dotenv==1.0.0
